Moved steamworks2017.Match into an inner-class of steamworks2017.MatchData.Match to avoid ambiguity with MatchSchedule.Match.
